AC1 - elim suppliers by zero score dimensions
